% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border.R
\name{border}
\alias{border}
\title{Borders and border colors}
\usage{
border(
  x,
  color = NULL,
  all = !is.null(color),
  top = NULL,
  right = NULL,
  bottom = NULL,
  left = NULL
)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{color}{One of the theme colors, defaults to \code{NULL}.}

\item{all}{One or more of \code{"top"}, \code{"t"}, \code{"right"}, \code{"r"}, \code{"bottom"},
\code{"b"}, or \code{"left"}, \code{"l"}, specifying which sides to add borders to,
defaults to \code{!is.null(color)}. \code{TRUE} and \code{FALSE} may be used as shorthands
for all sides or no sides, respectively.}

\item{top, right, bottom, left}{One of \code{TRUE} or \code{FALSE} specifying if a border
is added or removed to the tag element side, defaults to \code{NULL}, in which
case the argument is ignored.}
}
\description{
The \code{border()} functions adds or modifies tag element borders.
}
