% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border-width.R
\name{border_width}
\alias{border_width}
\title{Border width}
\usage{
border_width(x, width)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{width}{A number between \code{1} and
\code{5}.}
}
\description{
Border width description.
}
\examples{

.style \%>\%
  border_all() \%>\%
  border_width(3)

}
\seealso{
Other border utilities
\code{\link{border_all}()},
\code{\link{border_color}()}
}
\concept{border}
