% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display}
\alias{display}
\title{Display}
\usage{
display(x, ...)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{...}{Name-value pairs specifying \link{breakpoints}.}
}
\value{
An object of the same type as \code{x}.
}
\description{
The \code{display()} function adjusts how a tag element and its contents are
rendered.
}
\examples{

library(htmltools)

div(
  .style \%>\%
    display("flex") \%>\%
    flex_justify("center"),
  "Powerful stuff"
)

}
