% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed.R
\name{fixed}
\alias{fixed}
\title{Fixed position elements}
\usage{
fixed(x, position)
}
\arguments{
\item{position}{One of \code{"top"} or \code{"bottom"} specifying the fixed behavior of
an element.}

\item{.tag}{A tag element or \link{.style} pronoun.}
}
\description{
The \code{fixed()} function applies Bootstrap classes to fix elements to the top
or bottom of a page.
}
\seealso{
Other design utilities: 
\code{\link{background}()},
\code{\link{border}()},
\code{\link{display}()},
\code{\link{flex}()},
\code{\link{float}()},
\code{\link{font}()},
\code{\link{height}()},
\code{\link{margin}()},
\code{\link{overflow}()},
\code{\link{padding}()},
\code{\link{shadow}()},
\code{\link{width}()}
}
\concept{design utilities}
