% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex.R
\name{flex_display}
\alias{flex_display}
\alias{flex_direction}
\alias{flex_justify}
\alias{flex_align}
\alias{flex_gap}
\alias{flex_wrap}
\alias{flex_content}
\title{Flex}
\usage{
flex_display(x)

flex_direction(x, ...)

flex_justify(x, ...)

flex_align(x, ...)

flex_gap(x, ...)

flex_wrap(x, ...)

flex_content(x, ...)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{...}{Name-value pairs.}
}
\description{
The \verb{flex_*()} functions adjust the flexbox layout of an element. The flexbox
layout is incredibly powerful and allows centering of elements vertically and
horizontally, automatic adjustment of space between and around child
elements, and more.  To use flexbox make sure to include \code{flex_display()}
when styling an element. For finer control over an element's display see
\code{\link[=display]{display()}}.

Direct child elements of a flex box container are automatically considered
flex items and may be adjusted with the \verb{item_*()} functions, see
\code{\link[=item_align]{item_align()}}.
}
\examples{

library(htmltools)

div(
  .style \%>\%
    flex_display() \%>\%
    flex_justify("end"),
  div("Item 1"),
  div("Item 2"),
  div("Item 3")
)

}
