% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{font_size}
\alias{font_size}
\alias{font_weight}
\alias{font_style}
\title{Font}
\usage{
font_size(x, size)

font_weight(x, weight)

font_style(x, style)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{size}{One of \code{1}, \code{2}, \code{3}, \code{4}, \code{5}, or \code{6} specifying a font
size, defaults to \code{NULL}, in which case the argument is ignored. The sizes
follow the conventions of heading tags, so \code{1} is the largest font and \code{6}
the smallest.}

\item{weight}{One of \code{"bold"}, \code{"bolder"}, \code{"semibold"}, \code{"medium"}, \code{"normal"}, \code{"light"}, or \code{"lighter"} specifying the font
weight, defaults to \code{NULL}, in which case the argument is ignored.

If \code{"bolder"} or \code{"lighter"}, the font weight is changed relative to the
current font weight.}

\item{style}{One of \code{"italic"} or \code{"normal"} specifying the font
style, defaults to \code{NULL}, in which case the argument is ignored.}
}
\description{
The \verb{font_*()} functions adjust the size, weight, and style of a tag
element's font.
}
\examples{

library(htmltools)

p(
  .style \%>\%
    text_color("info") \%>\%
    font_weight("bold"),
  "The values above will include more value in v2.0"
)

}
