% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin.R
\name{margin}
\alias{margin}
\title{Margins}
\usage{
margin(
  x,
  all = NULL,
  top = NULL,
  right = NULL,
  bottom = NULL,
  left = NULL,
  ...
)
}
\arguments{
\item{all, top, right, bottom, left}{A \link{responsive} argument. One of \code{-5:5} or
\code{"auto"} specifying a margin for one or more sides of the tag element. 0
removes all outer space, 5 adds the most space, and negative values will
consume space pulling the element in that direction.}
}
\description{
The \code{margin} utilty changes the outer spacing of a tag element. The margin of
a tag element is the space outside and around the tag element, its border,
and its content. All arguments default to \code{NULL}, in which case they are
ignored.
}
\seealso{
Other design utilities: 
\code{\link{affix}()},
\code{\link{background}()},
\code{\link{border}()},
\code{\link{display}()},
\code{\link{flex}()},
\code{\link{float}()},
\code{\link{font}()},
\code{\link{height}()},
\code{\link{padding}()},
\code{\link{scroll}()},
\code{\link{shadow}()},
\code{\link{width}()}
}
\concept{design utilities}
