% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin.R
\name{margin_all}
\alias{margin_all}
\alias{margin_top}
\alias{margin_right}
\alias{margin_bottom}
\alias{margin_left}
\alias{margin_horizontal}
\alias{margin_vertical}
\title{Margins}
\usage{
margin_all(x, ...)

margin_top(x, ...)

margin_right(x, ...)

margin_bottom(x, ...)

margin_left(x, ...)

margin_horizontal(x, ...)

margin_vertical(x, ...)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{...}{One of \code{"-5"}, \code{"-4"}, \code{"-3"}, \code{"-2"}, \code{"-1"}, \code{0}, \code{1}, \code{2}, \code{3}, \code{4}, \code{5}, or \code{"auto"} specifying the margin
for one or all sides.

Use \link{breakpoints} to specify responsive values.}
}
\description{
The \code{margin()} function adjusts the outer spacing of a tag element. The
margin of a tag element is the space outside and around the tag element, its
border, and its content.
}
\examples{

library(htmltools)

div(
  .style \%>\%
    margin_left(3) \%>\%
    margin_right(3),
  "Mauris mollis tincidunt felis."
)

div(
  .style \%>\%
    margin_horizontal(3),
  "Nulla posuere."
)

div(
  .style \%>\%
    margin_right("auto"),
  "Sed id ligula quis est convallis tempor."
)

}
