% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overflow.R
\name{overflow}
\alias{overflow}
\title{Overflow}
\usage{
overflow(x, scroll = "auto")
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{scroll}{One of \code{"auto"} or \code{"hidden"} specifying if the content of the
element scrolls, defaults to \code{"auto"}. \code{TRUE} and \code{FALSE} may be used in
place of \code{"auto"} or \code{"hidden"}, respectively.}
}
\description{
The \code{overflow()} function specifies an element's content should scroll
instead of extending the height or width of the element. The height of the
element must be set.
}
\seealso{
Other design utilities: 
\code{\link{background}()},
\code{\link{border}()},
\code{\link{display}()},
\code{\link{fixed}()},
\code{\link{flex}()},
\code{\link{float}()},
\code{\link{font}()},
\code{\link{height}()},
\code{\link{margin}()},
\code{\link{padding}()},
\code{\link{shadow}()},
\code{\link{width}()}
}
\concept{design utilities}
