% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padding.R
\name{padding}
\alias{padding}
\title{Padding}
\usage{
padding(x, all = NULL, top = NULL, right = NULL, bottom = NULL, left = NULL)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{all}{A \link{responsive} argument.

One of \code{1:5} specifying a padding for all sides of the tag element,
defaults to \code{NULL}, in which case the argument is ignored. 0 removes all
inner space and 5 adds the most space.}

\item{top, right, bottom, left}{A \link{responsive} argument.

One of \code{1:5} specifying a padding for the element's respective side,
defaults to \code{NULL}, in which case the argument is ignored. 0 removes all
inner space and 5 adds the most space.}
}
\description{
The \code{padding()} function adjusts the inner spacing of a tag element. The
padding of a tag element is the space between the tag element's border and
its content or child elements.
}
