% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padding.R
\name{padding_all}
\alias{padding_all}
\alias{padding_top}
\alias{padding_right}
\alias{padding_bottom}
\alias{padding_left}
\alias{padding_horizontal}
\alias{padding_vertical}
\title{Add space into elements}
\usage{
padding_all(x, ...)

padding_top(x, ...)

padding_right(x, ...)

padding_bottom(x, ...)

padding_left(x, ...)

padding_horizontal(x, ...)

padding_vertical(x, ...)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{...}{One of the following,
\itemize{
\item \code{0}
\item \code{1}
\item \code{2}
\item \code{3}
\item \code{4}
\item \code{5}

Use \link{breakpoints} to specify responsive values.
}}
}
\description{
The \verb{padding_*()} functions adjust the inner spacing of a tag element. The
padding of a tag element is the space between the element's border and its
content or child elements.
}
\examples{

library(htmltools)

div(
  .style \%>\%
    margin_all(2) \%>\%
    padding_all(2) \%>\%
    border_color("primary") \%>\%
    background_color("light"),
  "This element has padding"
)

}
