% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{position_absolute}
\alias{position_absolute}
\alias{position_static}
\alias{position_relative}
\alias{position_fixed}
\alias{position_sticky}
\title{Positioning elements}
\usage{
position_absolute(x)

position_static(x)

position_relative(x)

position_fixed(x)

position_sticky(x)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}
}
\description{
The \verb{position_*()} functions adjust set the position of an element.
}
\examples{

.style \%>\%
  position_absolute() \%>\%
  position_top(50)

.style \%>\%
  position_relative() \%>\%
  position_left(50) \%>\%
  position_translate()

}
\seealso{
Other position: 
\code{\link{position_top}()},
\code{\link{position_translate}()}
}
\concept{position}
