% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{position_top}
\alias{position_top}
\alias{position_left}
\alias{position_bottom}
\alias{position_right}
\title{Arranging positioned elements}
\usage{
position_top(x, at)

position_left(x, at)

position_bottom(x, at)

position_right(x, at)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{at}{One of the following character strings,

\code{0}, \code{50}, or \code{100}}
}
\description{
Arrange elemnents when positioning then.
}
\seealso{
Other position: 
\code{\link{position_absolute}()},
\code{\link{position_translate}()}
}
\concept{position}
