% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{position_translate}
\alias{position_translate}
\title{Centering positioned elements}
\usage{
position_translate(x)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}
}
\description{
\code{position_translate()}
}
\seealso{
Other position: 
\code{\link{position_absolute}()},
\code{\link{position_top}()}
}
\concept{position}
