% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounded.R
\name{rounded_all}
\alias{rounded_all}
\alias{rounded_top}
\alias{rounded_right}
\alias{rounded_bottom}
\alias{rounded_left}
\title{Element corners}
\usage{
rounded_all(x, radius = 1)

rounded_top(x, radius = 1)

rounded_right(x, radius = 1)

rounded_bottom(x, radius = 1)

rounded_left(x, radius = 1)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{radius}{One of \code{0}, \code{1}, \code{2}, \code{3}, \code{4}, \code{5}, \code{"circle"}, or \code{"pill"}, defaults to
\code{1}. \code{0} will remove rounded corners.}
}
\description{
Round element corners or remove rounded corners.
}
