% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow.R
\name{shadow}
\alias{shadow}
\title{Shadows}
\usage{
shadow(x, size)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{size}{One of \code{"small"}, \code{"sm"}, \code{"medium"}, \code{"md"}, \code{"large"}, \code{"lg"}, or \code{"none"} specifying the amount
of shadow added.}
}
\description{
The \code{shadow()} function adjusts the box shadow of a tag element. Shadows help
distinguish elements or indicate interactivity.
}
