% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-alignment.R
\name{text_alignment}
\alias{text_alignment}
\title{Align text based on browser size}
\usage{
text_alignment(x, ...)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{...}{Name-value pairs. Names are \link{breakpoints}. Values are one or more
of \code{"left"}, \code{"right"}, or \code{"center"}.}
}
\description{
\code{text_alignment()} specifies how the text within a tag element is
aligned. Text may be aligned to the left, right, or center.
}
\examples{

.style \%>\%
  text_alignment("left")

.style \%>\%
 text_alignment("center")

}
\seealso{
Other text utilities: 
\code{\link{text_break}()},
\code{\link{text_color}()},
\code{\link{text_decoration}()},
\code{\link{text_selection}()},
\code{\link{text_transform}()},
\code{\link{text_wrap}()}
}
\concept{text utilities}
