% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-color.R
\name{text_color}
\alias{text_color}
\title{Emphasize text through color}
\usage{
text_color(x, color)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{color}{A character string specifying one of the following colors,
\itemize{
\item \code{"primary"}
\item \code{"secondary"}
\item \code{"success"}
\item \code{"danger"}
\item \code{"warning"}
\item \code{"info"}
\item \code{"light"}
\item \code{"dark"}
\item \code{"body"}
\item \code{"black"}
\item \code{"white"}
}}
}
\description{
The \code{text_color()} function adjusts the color of a tag element's text.
}
\examples{

.style \%>\%
  text_color("primary")

.style \%>\%
  text_color("dark") \%>\%
  border_color("dark")

}
\seealso{
Other text utilities: 
\code{\link{text_alignment}()},
\code{\link{text_break}()},
\code{\link{text_decoration}()},
\code{\link{text_selection}()},
\code{\link{text_transform}()},
\code{\link{text_wrap}()}
}
\concept{text utilities}
