% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-decoration.R
\name{text_decoration}
\alias{text_decoration}
\title{Text decoration}
\usage{
text_decoration(x, decoration)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{decoration}{One of the following character strings,
\itemize{
\item \code{"underline"}
\item \code{"strike"}
\item \code{"none"}
}}
}
\description{
Text decoration
}
\examples{

.style \%>\%
  text_decoration("strike")

.style \%>\%
  text_color("danger") \%>\%
  text_decoration("underline")

}
\seealso{
Other text utilities: 
\code{\link{text_alignment}()},
\code{\link{text_break}()},
\code{\link{text_color}()},
\code{\link{text_selection}()},
\code{\link{text_transform}()},
\code{\link{text_wrap}()}
}
\concept{text utilities}
