% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-transform.R
\name{text_transform}
\alias{text_transform}
\title{Text transformations}
\usage{
text_transform(x, transform)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{transform}{A character string.}
}
\description{
Text transformation
}
\examples{

.style \%>\%
  text_transform("lowercase")

.style \%>\%
  text_color("warning") \%>\%
  text_transform("uppercase")

}
\seealso{
Other text utilities: 
\code{\link{text_alignment}()},
\code{\link{text_break}()},
\code{\link{text_color}()},
\code{\link{text_decoration}()},
\code{\link{text_selection}()},
\code{\link{text_wrap}()}
}
\concept{text utilities}
