% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-wrap.R
\name{text_wrap}
\alias{text_wrap}
\title{Text wrap}
\usage{
text_wrap(x, wrap)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{wrap}{One of \code{TRUE} or \code{FALSE}.}
}
\description{
Text wrap
}
\examples{

.style \%>\%
  text_wrap(TRUE)

.style \%>\%
  text_wrap(FALSE)

}
\seealso{
Other text utilities: 
\code{\link{text_alignment}()},
\code{\link{text_break}()},
\code{\link{text_color}()},
\code{\link{text_decoration}()},
\code{\link{text_selection}()},
\code{\link{text_transform}()}
}
\concept{text utilities}
