% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visible.R
\name{visible}
\alias{visible}
\title{Element visibility}
\usage{
visible(x, value)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{value}{One of \code{TRUE} or \code{FALSE} specifying if the element is visible,
defaults to \code{TRUE}.}
}
\description{
The \code{visible()} function changes the visibility of a tag element. An
invisible element is both visually hidden and is also hidden from screen
readers.
}
\examples{

library(htmltools)

div("I am hidden") \%>\%
  visible(FALSE)

}
