% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-style.R
\name{with_style_pronoun}
\alias{with_style_pronoun}
\title{Style pronoun contexts}
\usage{
with_style_pronoun(pronoun, expr)
}
\arguments{
\item{pronoun}{A call to \code{style_pronoun()}.}

\item{expr}{An expression.}
}
\description{
The \code{with_style_pronoun()} allows overloading the \code{.style} pronoun.
}
\keyword{internal}
