% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{Inline alignment}
\usage{
align(.tag, align, ...)
}
\arguments{
\item{.tag}{A tag element or \code{.style} pronoun.}

\item{align}{One of \verb{r dQuote(align_values, q = FALSE)}}

\item{...}{Other arguments passed on to methods.}
}
\description{
The \code{align()} function adjusts the CSS \code{vertical-align} property of an
element. This property applies to \strong{inline} elements and may be used to
adjust the alignment of an image in a line of text or the contents of a table
cell.

For broader alignment purposes you may want to consider flex \code{display()} and
\code{flex()}.
}
