% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{Inline alignment}
\usage{
align(x, vertical)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{vertical}{One of
\Sexpr[results=rd,stage=render]{rd_quoted(align_values)}.}
}
\description{
The \code{align()} function adjusts the CSS \code{vertical-align} property of an
element. This property applies to \strong{inline} elements and may be used to
adjust the alignment of an image in a line of text or the contents of a table
cell.

For broader alignment purposes use flex box, see \code{display()} and \code{flex()}.
}
