% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{background}
\alias{background}
\title{Backgrounds}
\usage{
background(x, color)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{color}{One of \code{"blue"}, \code{"indigo"}, \code{"purple"}, \code{"red"}, \code{"orange"}, \code{"yellow"}, \code{"green"}, \code{"teal"}, \code{"cyan"}, \code{"white"}, or \code{"transparent"} specifying the
background color of the tag element.}
}
\description{
The \code{background()} function adjusts the background color of a tag element.
}
\section{Buttons}{
Use \code{background()} to modify shiny’s action buttons.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{actionButton("id", "Take action") \%>\%
  background("green")
}\if{html}{\out{</div>}}

\if{html}{\out{<button id="id" type="button" class="btn btn-default action-button cas-bg-green">}}Take action\if{html}{\out{</button>}}

With a couple other functions we can take our improvement a step
further.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{actionButton("id", "Take action") \%>\%
  background("green") \%>\%
  border("green") \%>\%
  shadow("small")
}\if{html}{\out{</div>}}

\if{html}{\out{<button id="id" type="button" class="btn btn-default action-button cas-bg-green cas-border cas-border-green cas-border-round-md cas-shadow-sm">}}Take action\if{html}{\out{</button>}}

Shiny’s download buttons include \code{...}, so we can use the \code{.style}
pronoun!

\if{html}{\out{<div class="sourceCode r">}}\preformatted{downloadButton(
  .style \%>\%
    background("white") \%>\%
    border("blue"),
  outputId = "dwnld",
  label = "Do a download",
  class = NULL
)
}\if{html}{\out{</div>}}
}

\section{Panels}{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{sidebarLayout(
  sidebarPanel(
    .style \%>\%
      background("blue"),
    "It's alive"
  ),
  mainPanel(
    .style \%>\%
      background("red"),
    "It's panel"
  )
)
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="row">
<div class="col-sm-4">
<form class="well cas-bg-blue" role="complementary">It's alive</form>
</div>
<div class="col-sm-8 cas-bg-red" role="main">It's panel</div>
</div>
}}
}

\section{Colors}{
\if{html}{\out{
<div class="cas-d-flex cas-flex-justify-between cas-flex-wrap">
<div class="cas-m-3">
<div class="cas-text-center">blue</div>
<div class="cas-p-5 cas-bg-blue"></div>
</div>
<div class="cas-m-3">
<div class="cas-text-center">indigo</div>
<div class="cas-p-5 cas-bg-indigo"></div>
</div>
<div class="cas-m-3">
<div class="cas-text-center">purple</div>
<div class="cas-p-5 cas-bg-purple"></div>
</div>
<div class="cas-m-3">
<div class="cas-text-center">red</div>
<div class="cas-p-5 cas-bg-red"></div>
</div>
<div class="cas-m-3">
<div class="cas-text-center">orange</div>
<div class="cas-p-5 cas-bg-orange"></div>
</div>
<div class="cas-m-3">
<div class="cas-text-center">yellow</div>
<div class="cas-p-5 cas-bg-yellow"></div>
</div>
<div class="cas-m-3">
<div class="cas-text-center">green</div>
<div class="cas-p-5 cas-bg-green"></div>
</div>
<div class="cas-m-3">
<div class="cas-text-center">teal</div>
<div class="cas-p-5 cas-bg-teal"></div>
</div>
<div class="cas-m-3">
<div class="cas-text-center">cyan</div>
<div class="cas-p-5 cas-bg-cyan"></div>
</div>
<div class="cas-m-3">
<div class="cas-text-center">white</div>
<div class="cas-p-5 cas-bg-white"></div>
</div>
<div class="cas-m-3">
<div class="cas-text-center">transparent</div>
<div class="cas-p-5 cas-bg-transparent"></div>
</div>
</div>
}}
}

\examples{

library(htmltools)

div(
  .style \%>\%
    background("white") \%>\%
    border("blue") \%>\%
    text("white"),
  "Nunc porta vulputate tellus.",
  "Suspendisse potenti."
)

}
