% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{background}
\alias{background}
\title{Backgrounds}
\usage{
background(x, color)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{color}{One of
\Sexpr[results=rd,stage=render]{rd_chr_lst(theme_colors)}.}
}
\description{
Use \code{background()} to modify the background color of a tag element.
}
