% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background-color.R
\name{background_color}
\alias{background_color}
\alias{subtle}
\title{Background color}
\usage{
background_color(x, color)

subtle(color)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{color}{One of \code{"primary"}, \code{"secondary"}, \code{"success"}, \code{"danger"}, \code{"warning"}, \code{"info"}, \code{"light"}, \code{"dark"}, \code{"body"}, \code{"black"}, \code{"white"}, or \code{"transparent"} specifying
the background color of the tag element.}
}
\description{
The \code{background_color()} function adjusts the background color of a tag
element.
}
\examples{

library(htmltools)

div(
  .style \%>\%
    background_color("white") \%>\%
    border_color("primary") \%>\%
    text_color("primary"),
  "Nunc porta vulputate tellus.",
  "Suspendisse potenti."
)

}
