% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border.R
\name{border}
\alias{border}
\title{Borders}
\usage{
border(x, color, sides = TRUE)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{color}{One of \code{"red"}, \code{"purple"}, \code{"indigo"}, \code{"blue"}, \code{"cyan"}, \code{"teal"}, \code{"green"}, \code{"yellow"}, \code{"amber"}, \code{"orange"}, or \code{"white"}.}

\item{sides}{One or more of \code{"top"}, \code{"t"}, \code{"right"}, \code{"r"}, \code{"bottom"}, \code{"b"}, \code{"left"}, or \code{"l"} specifying
which sides to add borders to, defaults to \code{TRUE}. \code{TRUE} and \code{FALSE} may
be used as shorthands for all sides or no sides, respectively.}
}
\description{
The \code{border()} function adjusts a tag element's borders.
}
