% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border.R
\name{border}
\alias{border}
\title{Borders}
\usage{
border(x, color, sides = TRUE, width = 1, round = "medium")
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{color}{One of \code{"blue"}, \code{"indigo"}, \code{"purple"}, \code{"red"}, \code{"orange"}, \code{"yellow"}, \code{"green"}, \code{"teal"}, \code{"cyan"}, or \code{"white"} specifying the border
color.}

\item{sides}{One or more of \code{"top"}, \code{"t"}, \code{"right"}, \code{"r"}, \code{"bottom"}, \code{"b"}, \code{"left"}, \code{"l"}, \code{"all"}, or \code{"none"} specifying
which sides to add borders to, defaults to \code{TRUE}. \code{TRUE} and \code{FALSE} may
be used as shorthand for all sides or no sides, respectively.}

\item{width}{One or more of \code{1}, \code{2}, \code{3}, \code{4}, or \code{5} specifying the width
of the element's border, defaults to \code{1}.}

\item{round}{One of \code{"small"}, \code{"sm"}, \code{"medium"}, \code{"md"}, \code{"large"}, \code{"lg"}, or \code{"none"} specifying how to round
the corners of the element, defaults to \code{"medium"}.}
}
\description{
The \code{border()} function adjusts a tag element's borders.
}
\section{Colors}{
\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="cas-d-flex">
}}
\if{html}{\out{
<div class="cas-p-4 cas-m-3 cas-border cas-border-blue cas-border-round-md">
}}


blue\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-p-4 cas-m-3 cas-border cas-border-indigo cas-border-round-md">
}}


indigo\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-p-4 cas-m-3 cas-border cas-border-purple cas-border-round-md">
}}


purple\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-p-4 cas-m-3 cas-border cas-border-red cas-border-round-md">
}}


red\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-p-4 cas-m-3 cas-border cas-border-orange cas-border-round-md">
}}


orange\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-p-4 cas-m-3 cas-border cas-border-yellow cas-border-round-md">
}}


yellow\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-p-4 cas-m-3 cas-border cas-border-green cas-border-round-md">
}}


green\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-p-4 cas-m-3 cas-border cas-border-teal cas-border-round-md">
}}


teal\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-p-4 cas-m-3 cas-border cas-border-cyan cas-border-round-md">
}}


cyan\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-p-4 cas-m-3 cas-border cas-border-white cas-border-round-md">
}}


white\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}
}

\section{Round}{
\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="cas-d-flex">
}}
\if{html}{\out{
<div class="cas-p-4 cas-m-3 cas-border cas-border-blue cas-border-round-sm">
}}


small\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-p-4 cas-m-3 cas-border cas-border-blue cas-border-round-sm">
}}


sm\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-p-4 cas-m-3 cas-border cas-border-blue cas-border-round-md">
}}


medium\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-p-4 cas-m-3 cas-border cas-border-blue cas-border-round-md">
}}


md\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-p-4 cas-m-3 cas-border cas-border-blue cas-border-round-lg">
}}


large\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-p-4 cas-m-3 cas-border cas-border-blue cas-border-round-lg">
}}


lg\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-p-4 cas-m-3 cas-border cas-border-blue cas-border-round-none">
}}


none\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}
}

\examples{

library(htmltools)

h3(
  .style \%>\%
    border("red", "bottom") \%>\%
    text("red"),
  "A bright, underlined heading!"
)

}
