% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border-color.R
\name{border_color}
\alias{border_color}
\title{Border colors}
\usage{
border_color(x, color)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{color}{A character string specifying one of the following colors,
\itemize{
\item \code{"primary"}
\item \code{"secondary"}
\item \code{"success"}
\item \code{"danger"}
\item \code{"warning"}
\item \code{"info"}
\item \code{"light"}
\item \code{"dark"}
\item \code{"black"}
\item \code{"white"}
}}
}
\description{
Border color description.
}
\seealso{
Other border utilities
\code{\link{border_all}()},
\code{\link{border_width}()}
}
\concept{border}
