% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border-width.R
\name{border_width}
\alias{border_width}
\title{Border width}
\usage{
border_width(x, width)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{width}{A number between \code{1} and
\code{5}.}
}
\value{
An object of the same type as \code{x}.
}
\description{
Adjust the boder width of a tag element.
}
\examples{

library(htmltools)

div(
  .style \%>\%
    border_all() \%>\%
    border_width(3) \%>\%
    border_color(theme_primary())
)

}
\seealso{
Other border utilities: 
\code{\link{border_color}()}
}
\concept{border utilities}
