% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascadess.R
\name{cascadess}
\alias{cascadess}
\title{Cascadess}
\usage{
cascadess()
}
\description{
Styles for htmltools tags.

For styles to be applied you must include a call to \code{cascadess()} in your
shiny application or htmltools tags.
}
\examples{

\dontrun{
library(shiny)

shinyApp(
  ui = list(
    cascadess(),
    div(
      .style \%>\%
        padding(3) \%>\%
        background("indigo") \%>\%
        font("light"),
      "Etiam laoreet quam sed arcu."
    )
  ),
  server = function(input, output) {

  }
)
}

}
