% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display}
\alias{display}
\title{Display}
\usage{
display(x, type)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{type}{A \link{responsive} argument.

One of \code{"inline"}, \code{"inline-block"}, \code{"block"}, \code{"grid"}, \code{"table"}, \code{"table-cell"}, \code{"table-row"}, \code{"flex"}, \code{"inline-flex"}, or \code{"none"}.}
}
\description{
The \code{display()} function adjusts how a tag element is rendered. For example,
to use the flex box layout the display must be \code{"flex"}.
}
\section{Block vs inline}{
\if{html}{\out{<div class="r">}}\preformatted{div(
  div(
    .style \%>\%
      border("blue"),
    "block"
  ),
  div(
    .style \%>\%
      border("blue"),
    "block"
  )
)
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div>
}}
\if{html}{\out{
<div class="cas-border-blue cas-border-all cas-border-round-md">
}}


block\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-border-blue cas-border-all cas-border-round-md">
}}


block\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}
\if{html}{\out{<div class="r">}}\preformatted{div(
  div(
    .style \%>\%
      border("blue") \%>\%
      display("inline"),
    "inline"
  ),
  div(
    .style \%>\%
      border("blue") \%>\%
      display("inline"),
    "inline"
  )
)
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div>
}}
\if{html}{\out{
<div class="cas-border-blue cas-border-all cas-border-round-md cas-display-inline">
}}


inline\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-border-blue cas-border-all cas-border-round-md cas-display-inline">
}}


inline\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}
}

\examples{

library(htmltools)

# When using flex make sure you specify the flex display.
div(
  .style \%>\%
    display("flex") \%>\%
    flex(justify = "center"),
  "Powerful stuff"
)

}
