% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display}
\alias{display}
\title{Display}
\usage{
display(x, type)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{type}{A \link{responsive} argument.

One of \code{"inline"}, \code{"inline-block"}, \code{"block"}, \code{"table"}, \code{"table-row"},
\code{"table-cell"}, \code{"flex"}, \code{"inline-flex"}, or \code{"none"}.}
}
\description{
The \code{display()} function adjusts how a tag element is rendered. All arguments
are responsive allowing you to hide elements on small screens or convert
elements from inline to block on large screens.
}
