% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed.R
\name{fixed}
\alias{fixed}
\title{Fixed position elements}
\usage{
fixed(x, position)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{position}{One of \code{"top"} or \code{"bottom"} specifying the where to fix the
element on the page.}
}
\description{
The \code{fixed()} function affixes an element to the top or bottom of the
page. Because of their nature fixed elements may cover up other elements on
the page. Use \code{\link[=padding]{padding()}} or \code{\link[=margin]{margin()}} to adjust any covered elements.
}
