% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex.R
\name{flex}
\alias{flex}
\title{Flex items}
\usage{
flex(x, align = "stretch", order = 1, grow = NULL, shrink = NULL)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{align}{A \link{responsive} argument.

One of \code{"auto"}, \code{"start"}, \code{"end"}, \code{"center"}, \code{"baseline"}, or
\code{"stretch"} specifying how to align the item on the cross axis, defaults
to \code{"stretch"}. Overrides the \code{\link[=flexbox]{flexbox()}} \code{align} argument.}

\item{order}{A \link{responsive} argument.

One of \Sexpr[results=rd,stage=render]{rd_num_lst(0:7)} specifying the
order of the item, defaults to \code{1}. Items of the same order are then sorted
by their source code order.}

\item{grow}{A \link{responsive} argument.

One of \code{TRUE} or \code{FALSE}, defaults to \code{NULL}, in which case the argument
is ignored.}

\item{shrink}{A \link{responsive} argument.

One of \code{TRUE} or \code{FALSE}, defaults to \code{NULL}, in which case the argument
is ignored.}
}
\description{
The \code{flex()} function adjusts a flex item. Unlike \code{\link[=flexbox]{flexbox()}}, which adjusts
the flex box layout through the flex container element, \code{flex()} is used to
change specific flex items. A flex item may be reordered, expanded, or
shrunk.
}
