% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/float.R
\name{float}
\alias{float}
\title{Floats}
\usage{
float(x, ...)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{...}{One of the following,
\itemize{
\item \code{"left"}
\item \code{"l"}
\item \code{"right"}
\item \code{"r"}
\item \code{"none"}

Use \link{breakpoints} to specify responsive values.
}}
}
\description{
The \code{float()} function places an element to the left or right side of its
parent element. Other text and inline elements wrap around floated elements.
Note, \code{float()} has no effect on flex items.
}
\examples{

library(htmltools)

div(
  div(
    .style \%>\%
      border_color("danger") \%>\%
      float("left"),
    "Danger!"
  ),
  div(
    "Nam a sapien.",
    "Phasellus neque orci, porta a, aliquet quis, semper a, massa.",
    "Phasellus lacus."
  )
)

}
