% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{font}
\alias{font}
\title{Font}
\usage{
font(x, color = NULL, weight = NULL, case = NULL, align = NULL)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{color}{One of \Sexpr[results=rd,stage=render]{rd_chr_lst(theme_colors)}
specifying the font color, defaults to \code{NULL}, in which case the argument is
ignored.}

\item{weight}{One of
\Sexpr[results=rd,stage=render]{rd_chr_lst(font_weights)} specifying the
font weight of the element's text, defaults to \code{NULL}.

If \code{"bolder"} or \code{"lighter"}, the font weight is changed relative to the
current font weight.}

\item{case}{One of \code{"upper"}, \code{"lower"}, or \code{"title"} specifying a
transformation of the tag element's text, default to \code{NULL}, in which case
the argument is ignored.}

\item{align}{A \link{responsive} argument.

One of \Sexpr[results=rd,stage=render]{rd_chr_lst(font_aligns)} specifying
the alignment of the tag element's text, defaults to \code{NULL}, in which case
the argument is ignored.}
}
\description{
The \code{font()} function adjusts the color, size, weight, case, or alignment of
a tag element's text.
}
