% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex.R
\name{item_align}
\alias{item_align}
\alias{item_fill}
\alias{item_grow}
\alias{item_shrink}
\alias{item_order}
\title{Flex items}
\usage{
item_align(x, ...)

item_fill(x, ...)

item_grow(x, ...)

item_shrink(x, ...)

item_order(x, ...)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{...}{Name-value pairs.}
}
\description{
The \verb{item_*()} functions adjust a flex item element. Unlike the \verb{flex_*()}
utilities, which adjust the flexbox layout of a parent flexbox container, the
\verb{item_*()} functions are applied to individual flex items, child elements of
a flexbox container. A flex item may be realigned, reordered, expanded, or
shrunk.
}
\examples{

library(htmltools)

div(
  .style \%>\%
    flex_display(),
  div(
    .style \%>\%
      item_order("last"),
    "First"
  ),
  div(
    "Second"
  ),
  div(
    "Third"
  )
)

}
