% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-style.R
\name{local_style}
\alias{local_style}
\alias{with_style}
\title{Style pronoun contexts}
\usage{
local_style(..., .env = caller_env())

with_style(.expr, ...)
}
\arguments{
\item{pronoun}{A call to \code{style_pronoun()}.}

\item{expr}{An expression.}
}
\description{
The \code{local_style()} establishes new prefixes and is used to overload the
\code{.style} pronoun's defaults. The \code{with_style()} function provides a different
approach to achieve the same utility.
}
\keyword{internal}
