% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin.R
\name{margin}
\alias{margin}
\title{Margins}
\usage{
margin(
  x,
  all = NULL,
  top = NULL,
  right = NULL,
  bottom = NULL,
  left = NULL,
  horizontal = NULL,
  vertical = NULL
)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{all}{A \link{responsive} argument.

One of \code{-5:5} or \code{"auto"} specifying a margin for all sides of the tag
element, defaults to \code{NULL}, in which case the argument is ignored. 0
removes all outer space, 5 adds the most space, and negative values will
consume space and pull the element in that direction.}

\item{top, right, bottom, left}{A \link{responsive} argument.

One of \code{-5:5} or \code{"auto"} specifying a margin for the respective side of
the tag element. 0 removes all outer space, 5 adds the most space, and
negative values will consume space and pull the element in that direction.}
}
\description{
The \code{margin()} function adjusts the outer spacing of a tag element. The
margin of a tag element is the space outside and around the tag element, its
border, and its content.
}
