% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overflow.R
\name{overflow}
\alias{overflow}
\title{Overflow}
\usage{
overflow(x, scroll)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{scroll}{One of \code{"auto"} or \code{"hidden"} specifying if the content of the
element scrolls. \code{TRUE} and \code{FALSE} may be used in place of \code{"auto"} or
\code{"hidden"}, respectively.}
}
\description{
The \code{overflow()} function adjust how an element's content scrolls. Scrolling
an element's contents may be helpful to prevent child elements from extending
the height or width of the element. The height of the element must be set.
}
