% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overflow.R
\name{overflow}
\alias{overflow}
\title{Overflow}
\usage{
overflow(x, scroll)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{scroll}{One of \code{"auto"}, \code{"hidden"}, \code{"visible"}, or \code{"scroll"} specifying how the
content of the element scrolls. \code{TRUE} and \code{FALSE} may be used in place of
\code{"scroll"} or \code{"hidden"}, respectively.}
}
\description{
The \code{overflow()} function adjust how an element's content scrolls. Scrolling
an element's contents may be helpful to prevent child elements from extending
the width or height of the element.
}
\examples{

library(htmltools)

div(
  .style \%>\%
    width(25) \%>\%
    overflow("hidden"),
  "Nullam libero mauris, consequat quis, varius et, dictum id, arcu."
)

}
