% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overflow.R
\name{overflow}
\alias{overflow}
\title{Overflow}
\usage{
overflow(x, scroll = "auto")
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{scroll}{One of \code{"auto"} or \code{"hidden"} specifying if the content of the
element scrolls, defaults to \code{"auto"}. \code{TRUE} and \code{FALSE} may be used in
place of \code{"auto"} or \code{"hidden"}, respectively.}
}
\description{
The \code{overflow()} function specifies an element's content should scroll
instead of extending the height or width of the element. The height of the
element must be set.
}
