% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-offset.R
\name{position_top}
\alias{position_top}
\alias{position_left}
\alias{position_bottom}
\alias{position_right}
\title{Position offset}
\usage{
position_top(x, offset)

position_left(x, offset)

position_bottom(x, offset)

position_right(x, offset)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{offset}{A number specifying a percent. One of,

\code{0}, \code{50}, or \code{100}}
}
\value{
An object of the same type as \code{x}.
}
\description{
Use \verb{position_<side>()} to adjust the position offset of a tag element.
}
\examples{

library(htmltools)

div(
  .style \%>\%
    position("absolute") \%>\%
    position_right(0)
)

}
\seealso{
Other position: 
\code{\link{position_centered}()},
\code{\link{position}()}
}
\concept{position}
