% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius.R
\name{radius}
\alias{radius}
\title{Border radius}
\usage{
radius(x, round, sides = TRUE)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{round}{One of \code{"small"}, \code{"sm"}, \code{"medium"}, \code{"md"}, \code{"large"}, \code{"lg"}, \code{"circle"}, \code{"pill"}, or \code{"none"}.}

\item{sides}{One or more of \code{"top"}, \code{"t"}, \code{"right"}, \code{"r"}, \code{"bottom"}, \code{"b"}, \code{"left"}, \code{"l"}, or \code{"all"} specifying
which sides' corners to round, defaults to \code{TRUE}. \code{TRUE} and \code{FALSE} may
be used as shorthands for all sides or no sides, respectively.}
}
\description{
The \code{radius()} function adjusts the border radius of a tag element.
}
