% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounded.R
\name{rounded_all}
\alias{rounded_all}
\alias{rounded_top}
\alias{rounded_right}
\alias{rounded_bottom}
\alias{rounded_left}
\title{Rounding element corners}
\usage{
rounded_all(x, radius = 1)

rounded_top(x, radius = 1)

rounded_right(x, radius = 1)

rounded_bottom(x, radius = 1)

rounded_left(x, radius = 1)
}
\arguments{
\item{x}{"r param_subject()"}

\item{radius}{Radius param.}
}
\description{
Description of "rounded()".
}
