% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow.R
\name{shadow}
\alias{shadow}
\title{Shadows}
\usage{
shadow(x, size = "medium")
}
\arguments{
\item{size}{One of \code{"none"}, \code{"small"}, \code{"medium"}, or \code{"large"} specifying
the amount of shadow added, defaults to \code{"medium"}.}
}
\description{
The \code{shadow} utility applies a shadow to a tag element. By default many
elements include a shadow to help distinguish them. Use \code{"none"} to remove an
element's shadow.
}
\seealso{
Other design utilities: 
\code{\link{affix}()},
\code{\link{background}()},
\code{\link{border}()},
\code{\link{display}()},
\code{\link{flex}()},
\code{\link{float}()},
\code{\link{font}()},
\code{\link{height}()},
\code{\link{margin}()},
\code{\link{padding}()},
\code{\link{scroll}()},
\code{\link{width}()}
}
\concept{design utilities}
