% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow.R
\name{shadow}
\alias{shadow}
\title{Shadows}
\usage{
shadow(x, size)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{size}{One of \code{"small"}, \code{"sm"}, \code{"medium"}, \code{"md"}, \code{"large"}, \code{"lg"}, or \code{"none"} specifying the amount
of shadow added.}
}
\description{
The \code{shadow()} function adjusts the box shadow of a tag element. Shadows help
distinguish elements or indicate interactivity.
}
\section{Sizes}{
\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="cas-display-flex">
}}
\if{html}{\out{
<div class="cas-margin-all-3 cas-padding-all-4 cas-shadow-sm">
}}


small\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-margin-all-3 cas-padding-all-4 cas-shadow-sm">
}}


sm\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-margin-all-3 cas-padding-all-4 cas-shadow-md">
}}


medium\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-margin-all-3 cas-padding-all-4 cas-shadow-md">
}}


md\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-margin-all-3 cas-padding-all-4 cas-shadow-lg">
}}


large\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-margin-all-3 cas-padding-all-4 cas-shadow-lg">
}}


lg\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="cas-margin-all-3 cas-padding-all-4 cas-shadow-none">
}}


none\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}
}

