% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-pronoun.R
\docType{data}
\name{style-pronoun}
\alias{style-pronoun}
\alias{.style}
\title{Style pronoun}
\usage{
.style
}
\description{
The \code{.style} pronoun allows you to define styles for a tag element within the
context of the element. Without the \code{.style} pronoun tag styles are applied
outside and after constructing a tag element.

\if{html}{\out{<div class="sourceCode R">}}\preformatted{div("") \%>\%
  background_color("primary") \%>\%
  display("flex") \%>\%
  flex_justify("between")
}\if{html}{\out{</div>}}

However, once the content of a tag element grows to more than a few lines,
associating the element's styles with the element becomes less and less
intuitive. In these situations, make use of the \code{.style} pronoun.

\if{html}{\out{<div class="sourceCode R">}}\preformatted{div(
  .style \%>\%
    border_color("primary") \%>\%
    text_color("primary"),
  p("First paragraph"),
  p("Second paragraph")
)
}\if{html}{\out{</div>}}
}
\section{Prefixing}{


Complex components such as \code{shiny::radioButtons()} or
\code{yonder::listGroupInput()} may need a non-standard prefix for the CSS
classes applied by cascadess' functions.
}

\keyword{datasets}
