% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-break.R
\name{text_break}
\alias{text_break}
\title{Text break}
\usage{
text_break(x)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}
}
\description{
Text break.
}
\examples{

.style \%>\%
  text_break()

}
\seealso{
Other text utilities: 
\code{\link{text_alignment}()},
\code{\link{text_color}()},
\code{\link{text_decoration}()},
\code{\link{text_selection}()},
\code{\link{text_transform}()},
\code{\link{text_wrap}()}
}
\concept{text utilities}
